move primary tab to lib utils
