+++ app/assets/javascripts/discourse/app/lib/utilities.js
@@ -769,3 +769,19 @@ export function cleanNullQueryParams(params) {
 export function getElement(node) {
   return node.nodeType === Node.TEXT_NODE ? node.parentElement : node;
 }
+
+export function isPrimaryTab() {
+  return new Promise((resolve) => {
+    if (capabilities.supportsServiceWorker) {
+      navigator.serviceWorker.addEventListener("message", (event) => {
+        resolve(event.data.primaryTab);
+      });
+
+      navigator.serviceWorker.ready.then((registration) => {
+        registration.active.postMessage({ action: "primaryTab" });
+      });
+    } else {
+      resolve(true);
+    }
+  });
+}
